+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/storage/SessionStorageTest.kt
@@ -452,14 +452,20 @@ class SessionStorageTest {
         json.put("url", "testUrl")
         json.put("source", Source.NEW_TAB.name)
         json.put("parentUuid", "")
-        assertEquals(Source.NEW_TAB, deserializeSession(json, restoreId = true, restoreParentId = true).source)
+        assertEquals(
+            Source.RESTORED,
+            deserializeSession(json, restoreId = true, restoreParentId = true).source
+        )
 
         val jsonInvalid = JSONObject()
         jsonInvalid.put("uuid", "testId")
         jsonInvalid.put("url", "testUrl")
         jsonInvalid.put("source", "invalidSource")
         jsonInvalid.put("parentUuid", "")
-        assertEquals(Source.NONE, deserializeSession(jsonInvalid, restoreId = true, restoreParentId = true).source)
+        assertEquals(
+            Source.RESTORED,
+            deserializeSession(jsonInvalid, restoreId = true, restoreParentId = true).source
+        )
     }
 
     @Suppress("UNCHECKED_CAST")